Rename RetryEnterPresenceWorker, call from AddTrackableWorker
ably.connect no longer enters presence, so we always want to enter
presence when a connection is made and channel attached.

Repurpose RetryEnterPresenceWorker to be a general presence entering worker,
post it after AddTrackableWorker on all occaisions.
